# src/main/java/com/example/nosqlcapstone/model/StatusDto.java
package com.example.nosqlcapstone.model;

import lombok.AllArgsConstructor;
import lombok.EqualsAndHashCode;
import lombok.Getter;
import lombok.Setter;

@Setter
@Getter
@AllArgsConstructor
@EqualsAndHashCode
public class StatusDto {

    private String status;

}